peakfit / pipeline / fit-grain: memory-safety and FF->PF seeding fixes
- peakfit/pool.py: split LM batch capacity into independent device-batch
  and host-residency limits (a 250GB host was OOM-killed while the old
  code reported 8GiB free), cap solve-batch at 256 (magma_spotrs_batched
  faults at large counts), and bisect-on-OOM as the backstop
- peakfit/lm.py: cap torch.compile CUDA-graph variants at 32 — shape-diverse
  data pinned ~41GB in never-evicted graph pools and OOM'd
- peakfit/jacobian_triton.py: int64 region index (int32 stride overflow
  past b>=68 wrote J to a garbage address -> illegal memory access)
- fit_grain/scan_driver.py: thread the per-voxel scan-refine loop
  (reader-lock on shared handles, torch pinned to 1 intra-op thread)
- pipeline seeding: accept both GrainID/ID header spellings, and wire a
  GrainsFile line into the c-omp indexer paramstest so seeded PF isn't
  silently run as a full orientation grid

+++ packages/midas_peakfit/midas_peakfit/jacobian_triton.py
@@ -50,7 +50,16 @@ if _TRITON_OK:
         block handles one pixel; computes residual + full Jacobian row.
         N is loaded into registers per thread (small for typical P).
         """
-        b = tl.program_id(0)
+        # int64 for the region index: it multiplies the LARGEST stride
+        # (stride_JB = M*N). For dense regions batched together — e.g. B=125
+        # regions of n_peaks=400 (N=3201) x M=10000 → stride_JB=3.2e7 — the
+        # product b*stride_JB exceeds int32 (2^31) once b>=68, wrapping to a
+        # negative offset and writing J to a garbage address → intermittent
+        # "illegal memory access". It only bites when 68+ of the biggest
+        # regions land in one batch, which is why sparse data and small
+        # isolation batches never triggered it. Cast so every b-derived
+        # address is 64-bit.
+        b = tl.program_id(0).to(tl.int64)
         m_block = tl.program_id(1)
 
         m_offsets = m_block * BLOCK_M + tl.arange(0, BLOCK_M)
